Role model and modifications to the UserData model

+++ src/test/java/com/sysco/ordermanager/domain/repository/RestaurantRepositoryTest.java
@@ -33,7 +33,7 @@ public class RestaurantRepositoryTest {
     public void createRestaurant(){
 
         UserData userData = new UserData(
-                1,
+                (long)1,
                 "vibodha",
                 "123"
         );
