+++ src/test/java/com/sysco/ordermanager/service/RestaurantServiceImpTest.java
@@ -55,7 +55,7 @@ public class RestaurantServiceImpTest {
     @Before
     public void setup(){
         UserData userData = new UserData(
-                1,
+                (long)1,
                 "vibodha",
                 "123"
         );
